// src/interfaces/contact.interface.ts
export interface ContactData {
    id?: number;
    invitationId?: number;
    isActive?: boolean;
    order?: number;
    groomContact?: string;
    brideContact?: string;
    groomFatherContact?: string;
    groomMotherContact?: string;
    brideFatherContact?: string; 
    brideMotherContact?: string;
};